requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pinecone-client==3.0.0
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
pypdf2==3.0.1
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
